update project name from 'com.example.zoopclientesample' to 'com.zoop.zoopandroidsample'
