+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.example.zoopclientsample"
+        applicationId "com.zoop.zoopandroidsample"
         minSdkVersion 18
         targetSdkVersion 29
         versionCode 1
