--- app/src/main/java/com/example/zoopclientsample/ToastHelper.kt
@@ -1,16 +1,0 @@
-package com.example.zoopclientsample
-
-import android.content.Context
-import android.widget.Toast
-
-class ToastHelper (val context: Context) {
-
-    fun showToast(sMessage: String) {
-            Toast.makeText(
-                context,
-                sMessage,
-                Toast.LENGTH_SHORT
-            ).show()
-    }
-
-}
