--- app/src/main/java/com/example/zoopclientsample/TransactionStatus.kt
@@ -1,5 +1,0 @@
-package com.example.zoopclientsample
-
-enum class TransactionStatus {
-    READY, PROCESSING, FINISHED, ERROR
-}
