--- app/src/main/java/com/example/zoopclientsample/adapter/TerminalAdapterListener.kt
@@ -1,7 +1,0 @@
-package com.example.zoopclientsample.adapter
-
-import com.example.zoopclientsample.model.TerminalModel
-
-interface TerminalAdapterListener {
-    fun terminalModelItemOnClick(model: TerminalModel?, position: Int)
-}
