--- app/src/main/java/com/example/zoopclientsample/api/RetrofitInstance.kt
@@ -1,19 +1,0 @@
-package com.example.zoopclientsample.api
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitInstance {
-    private var retrofit: Retrofit? = null
-    private const val BASE_URL = "https://api-beta.zoop.ws"
-    val retrofitInstance: Retrofit?
-        get() {
-            if (retrofit == null) {
-                retrofit = Retrofit.Builder()
-                    .baseUrl(BASE_URL)
-                    .addConverterFactory(GsonConverterFactory.create())
-                    .build()
-            }
-            return retrofit
-        }
-}
