+++ app/src/main/java/com/zoop/zoopandroidsample/Constants.kt
@@ -1,0 +1,21 @@
+package com.zoop.zoopandroidsample
+
+object Constants {
+
+    const val CREDIT = "CREDIT"
+
+    const val DEBIT = "DEBIT"
+
+    const val USERNAME = "USERNAME"
+
+    const val PASSWORD = "PASSWORD"
+
+    const val FIRST_NAME = "FIRST_NAME"
+
+    const val LAST_NAME = "LAST_NAME"
+
+    const val USER_TOKEN = "USER_TOKEN"
+
+    const val SELLER_ID = "SELLER_ID"
+
+}
