+++ app/src/main/java/com/zoop/zoopandroidsample/Preferences.kt
@@ -1,0 +1,23 @@
+package com.zoop.zoopandroidsample
+
+import android.content.Context
+import android.content.SharedPreferences
+
+class Preferences(context: Context) {
+
+    private val mSharedPreferences: SharedPreferences =
+        context.getSharedPreferences("userData", Context.MODE_PRIVATE)
+
+    fun storeString(key: String?, value: String?) {
+        mSharedPreferences.edit().putString(key, value).apply()
+    }
+
+    fun getStoredString(key: String?): String? {
+        return mSharedPreferences.getString(key, "")
+    }
+
+    fun clean() {
+        mSharedPreferences.edit().clear().apply()
+    }
+
+}
