+++ app/src/main/java/com/zoop/zoopandroidsample/TransactionStatus.kt
@@ -1,0 +1,5 @@
+package com.zoop.zoopandroidsample
+
+enum class TransactionStatus {
+    READY, PROCESSING, FINISHED, ERROR
+}
