+++ app/src/main/java/com/zoop/zoopandroidsample/adapter/TerminalAdapterListener.kt
@@ -1,0 +1,7 @@
+package com.zoop.zoopandroidsample.adapter
+
+import com.zoop.zoopandroidsample.model.TerminalModel
+
+interface TerminalAdapterListener {
+    fun terminalModelItemOnClick(model: TerminalModel?, position: Int)
+}
