+++ app/src/main/java/com/zoop/zoopandroidsample/api/LoginRequest.kt
@@ -1,0 +1,9 @@
+package com.zoop.zoopandroidsample.api
+
+import com.google.gson.annotations.SerializedName
+
+class LoginRequest(
+    @SerializedName("username") var username: String,
+    @SerializedName("password") var password: String,
+    @SerializedName("persist") var persist: Boolean
+)
