+++ app/src/main/res/layout/activity_receipt.xml
@@ -34,7 +34,7 @@
                 android:src="@drawable/logo_zoop_receipt"
                 android:contentDescription="@string/logo_zoop_receipt" />
 
-            <com.example.zoopclientsample.ui.AutoResizeTextView
+            <com.zoop.zoopandroidsample.ui.AutoResizeTextView
                 android:id="@+id/autoResizeTextViewPrintReceipt"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
